requirements.txt
# 深度学习框架
torch>=1.12.0
torchaudio>=0.12.0
torchvision>=0.13.0
transformers>=4.20.0
pytorch-crf>=0.7.2

# 科学计算
numpy>=1.21.0
scipy>=1.8.0
pandas>=1.4.0
scikit-learn>=1.1.0

# 音频处理
librosa>=0.9.0
soundfile>=0.10.0
speechbrain>=0.5.0
openai-whisper>=20231117

# EEG信号处理
mne>=1.0.0
pyedflib>=0.1.0

# 自然语言处理
nltk>=3.7
spacy>=3.4.0
textstat>=0.7.0
jieba>=0.42.1
datasets>=2.16.0
huggingface_hub>=0.20.0

# 可视化和实验管理
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
wandb>=0.12.0

# 配置管理
pyyaml>=6.0
omegaconf>=2.2.0

# 工具库
tqdm>=4.64.0
click>=8.1.0
rich>=12.0.0
openpyxl>=3.0.0  # Excel文件读写

# 视频下载和处理
you-get>=0.4.1650  # 视频下载工具
ffmpeg-python>=0.2.0  # FFmpeg Python绑定

# 测试
pytest>=7.0.0
pytest-cov>=3.0.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=7.7.0 